upfate server

+++ server/services/review.service.ts
@@ -14,7 +14,7 @@ export const createReviewService = asyncHandler(
     const { productId } = req.params;
 
     const userId = req.user?._id;
-    const existingReview = await ReviewModel.findOne({ userId });
+    const existingReview = await ReviewModel.findOne({ userId, productId });
 
     if (!productId) throw new BadRequestError("Product ID is required");
     const { rating, comment } = req.body;
@@ -51,14 +51,14 @@ export const getReviewsService = asyncHandler(
     next: NextFunction
   ) => {
     const { productId } = req.params;
-    const { page, limit, sort } = req.query;
+    const { page, limit, sort } = req.body;
     const skip = (Number(page) - 1) * Number(limit);
 
     if (!productId) throw new BadRequestError("Product ID is required");
     const reviews = await ReviewModel.find({ productId })
       .skip(skip)
       .limit(Number(limit))
-      .sort({ rating: sort === 'asc' ? 1 : -1 });
+      .sort({ rating: sort === "asc" ? 1 : -1 }).populate({path: 'userId', select: 'name email'});
     const total = await ReviewModel.countDocuments({ productId });
     res.status(200).json({
       status: "success",
@@ -75,19 +75,38 @@ export const getReviewsService = asyncHandler(
   }
 );
 
-
-// export const updateReviewService = asyncHandler(
-//   async (
-//     req: AuthenticatedRequestBody<IReviewRequest>,
-//     res: Response,
-//     next: NextFunction
-//   ) => {
-//     const { reviewId } = req.params;
-//     const userId = req.user?._id;
-//     const review = await ReviewModel.findOne({ _id: reviewId, userId });
-//   }
-// );
-
+export const updateReviewService = asyncHandler(
+  async (
+    req: AuthenticatedRequestBody<IReviewRequest>,
+    res: Response,
+    next: NextFunction
+  ) => {
+    const { reviewId } = req.params;
+    const userId = req.user?._id;
+    const review = await ReviewModel.findOne({ _id: reviewId });
+    if (!review) {
+      throw new BadRequestError("Review not found");
+    }
+    if (review.userId.toString() !== userId?.toString()) {
+      console.log("review.userId", review.userId);
+      console.log("userId", userId);
+      throw new BadRequestError("You are not allowed to update this review");
+    }
+    const { rating, comment } = req.body;
+    if (!rating || !comment) {
+      throw new BadRequestError("Rating and comment are required");
+    }
+    const updatedReview = await ReviewModel.findByIdAndUpdate(
+      { _id: reviewId },
+      { rating, comment },
+      { new: true }
+    );
+    res.status(200).json({
+      status: "success",
+      data: updatedReview,
+    });
+  }
+);
 
 export const deleteReviewService = asyncHandler(
   async (
@@ -97,9 +116,13 @@ export const deleteReviewService = asyncHandler(
   ) => {
     const { reviewId } = req.params;
     const userId = req.user?._id;
-    const review = await ReviewModel.findOne({ _id: reviewId, userId });
+    const isAdmin = req.user?.role === "admin";
+    const review = await ReviewModel.findOne({ _id: reviewId });
     if (!review) {
       throw new BadRequestError("Review not found");
+    }
+    if (review.userId.toString() !== userId?.toString() && !isAdmin) {
+      throw new BadRequestError("You are not allowed to delete this review");
     }
     await ReviewModel.deleteOne({ _id: reviewId });
     res.status(200).json({
